Attempting to make `unknown` architecture compile.
