+++ src/arch/common/intrin/cast.rs
