+++ src/arch/common/intrin/destride.rs
