+++ src/arch/common/intrin/eq.rs
