+++ src/arch/common/intrin/macros.rs
