+++ src/arch/common/intrin/merge.rs
