+++ src/arch/common/intrin/popcnt.rs
