+++ src/arch/common/intrin/sum.rs
