+++ src/arch/common/intrin/swizzle.rs
