+++ src/arch/common/mod.rs
@@ -1,2 +1,2 @@
 #[macro_use] pub mod intrin;
-pub mod vecs;
+#[macro_use] pub mod vecs;
